refactored reducers components to separate files
